modules/requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
GitPython==3.1.40
PyYAML==6.0.1
APScheduler==3.10.4
requests==2.31.0
python-engineio==4.8.0
python-socketio==5.10.0
eventlet==0.33.3 